Replaced entire AdventureGame folder in Repo
I had to do this because I could not get any of the scripts to work for player movement or collision, and I believe it's because I changed the unity editor version when I used the lab computer, so now I'll stick to using the original editor version on my laptop only. Unfortunately this also means I'll have to start this project over from scratch, and stay away from the lab computers.